backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic>=2.5.0
email-validator==2.1.0
python-dotenv>=1.0.0
transformers==4.44.0
torch>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0